test-runner - Use jq to extract pipeline name regardess of its location
Change 26504 allowed to have the name field from the pipeline not only on the first line. However, it broke cases where the pipeline includes more than one 'name' field, that may be used within other variables declared.

With jq via yaml2json, we ensure we are extracting the top-level 'name' variable, which is the pipeline name.

Change-Id: I67bbaeeaea8c132c83aee172ea5ad09cc30179da