dci-queue | Add -f option to remove-resource command
Currently, if we try to remove a resource that does not exist, it is added to the reserved resources in the queue. And if we want to delete it (or delete a resource that we have added by mistake), we need to remove the files created under the dci-queue directories.

This change allows the following:

- Add -f argument to force the deletion of an existing resource in the queue
- Do not remove a resource from the queue if it does not exist
- Unit tests for these

Change-Id: I1a826a5620cac38a6c86b1f9f709f48ac2de08ad