Update matching string for dci_job cmd
The dci_job command parses the dci-queue logs to gather the job name and the
job id. But the logs could look different if the task is using
dci-check-change or dci-pipeline. The latter was modified some time ago
and the string expected by dci_job is gone. Updating the regex to find
the job name and id info.
Additionally, we include all job IDs even if there was a fallback.

Change-Id: I8c764f83acd490f338c5b51e9e30574b6c579518