dci-pipeline-helper | retrieve OCP topic if KUBECONFIG is set
In the current dci-pipeline-helper, if KUBECONFIG is set in the execution environment where dci-pipeline is executed (i.e. exported beforehand), this check:

if ! [ -r "$KUBECONFIG" ]; then

is not passed, then all the logic that retrieves the OCP topic from the running cluster is not executed, and if not defined the topic, execution fails with this failure:

Traceback (most recent call last):
  File "/usr/bin/dci-pipeline", line 11, in <module>
    load_entry_point('dci-pipeline==0.8.2', 'console_scripts', 'dci-pipeline')()
  File "/usr/lib/python3.6/site-packages/dcipipeline/main.py", line 1250, in main
    options,
  File "/usr/lib/python3.6/site-packages/dcipipeline/main.py", line 1153, in run_jobdefs
    pipeline_id=options["pipeline_id"],
  File "/usr/lib/python3.6/site-packages/dcipipeline/main.py", line 574, in schedule_job
    jobdef["topic"],
KeyError: 'topic'

This change allows to retrieve the OCP topic also in the case that KUBECONFIG is provided as exported variable. Some extra code is added to keep compatibility with the case of not providing KUBECONFIG and retrieving it from inventory or pipeline file

Change-Id: Ie1ee3aa5e4d8b994691b3e90bb9826a38727e77a