Report success status in dci-pipeline execution logs
Currently, test-runner reports if a job is failing or not by checking a regex, looking for "Jobdef ... status ..." chain, but this is only reported for failed jobs. If we have a pipeline where the last job fails but the previous ones passed, as this info is not printed for passed jobs, it will take the status of the last job, meaning that all jobs will be reported as failed, but it's not correct.

Let's print the success status for passed jobs, so that test-runner can interpret it.

Change-Id: I7ff1901adc7155cfa5095bcb8494810d7b37a418