test-runner extends the list of tags instead of overriding it
For every pipeline stage it runs, the test-runner script sets the
dci_tags variable to a list including the "debug" tag and the change ID.

This is a problem with custom tags are provided from the command line or
the CR description, since the tags set by test-runner will override the
custom ones.

This change allows test-runner to extend the list of custom tags, if it
exists, instead of replacing it.

Change-Id: I47a8f96de5f267fedc627ee0c91d5af881ed6043