Update regular expression for feeding ZTP_NAMES list
Right now, all jobs that contains "ztp-spoke" keyword in acm_cluster_type are included under ZTP_NAMES list, then the code to update the gitops branch for the change is executed.
However, this is not needed for "ztp-spoke-clusterinstance" installation type.
Updating the regex to only apply this feature to "ztp-spoke" installation type.

Change-Id: Ieb26696cf42e6a0531a8ff99c38035a0a4a21b51